implemented runtime and exe creation for windows
