+++ em-mode.el
@@ -23,7 +23,7 @@
     table))
 
 (defun em-types ()
-  '("void" "bool" "char" "int" "float" "string"))
+  '("void" "bool" "char" "int" "u32" "u64" "s32" "s64" "float" "string"))
 
 (defun em-keywords ()
   '("if" "else" "for" "while" "return" "break" "continue" "true" "false"))
